# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/problems/{id}/assign": {
      "post": {
        "operationId": "DeliveryController_assign_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignOfficerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Assign or reassign the delivery officer",
        "tags": [
          "delivery"
        ]
      }
    },
    "/api/v1/problems/{id}/project/progress": {
      "post": {
        "operationId": "DeliveryController_progress_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update project progress",
        "tags": [
          "delivery"
        ]
      }
    },
    "/api/v1/problems/{id}/project/complete": {
      "post": {
        "description": "Does not resolve the problem. It raises the verification request addressed to the reporters; they decide whether it was fixed.",
        "operationId": "DeliveryController_complete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark the project complete and ask the citizens who reported it",
        "tags": [
          "delivery"
        ]
      }
    },
    "/api/v1/automations/{id}": {
      "patch": {
        "operationId": "WorkspaceController_toggle_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutomationToggleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Enable or pause an automation",
        "tags": [
          "delivery"
        ]
      }
    },
    "/api/v1/alerts/{id}/read": {
      "post": {
        "operationId": "WorkspaceController_read_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mark one of your own alerts read",
        "tags": [
          "delivery"
        ]
      }
    },
...
  },
  "info": {
    "title": "Jan Setu API",
    "description": "Generated — do not edit by hand. Run `npm run openapi:gen`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AssignOfficerDto": {
        "type": "object",
        "properties": {
          "officerId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          }
        },
        "required": [
          "officerId"
        ]
      },
      "ProjectProgressDto": {
        "type": "object",
        "properties": {
          "progress": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 100,
            "exclusiveMaximum": false
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "progress"
        ]
      },
      "AutomationToggleDto": {
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}